Now users can not interact with someone else's data, and admin and moderators  can. Api as enum added. Id was added to UserDto.cs.

+++ Diary.Api/Controllers/ReportController.cs
@@ -1,9 +1,16 @@
+using System.Security.Claims;
 using Asp.Versioning;
+using Diary.Application.Services;
 using Diary.Domain.Dto.Report;
+using Diary.Domain.Entity;
+using Diary.Domain.Enum;
+using Diary.Domain.Interfaces.Repositories;
 using Diary.Domain.Interfaces.Services;
 using Diary.Domain.Result;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace Diary.Api.Controllers;
 
@@ -14,6 +21,7 @@ namespace Diary.Api.Controllers;
 /// <response code="400">If report was not created/deleted/updated/received</response>
 /// <response code="500">If internal server error occured</response>
 /// <response code="401">If user is unauthorized</response>
+/// <response code="403">If user is forbidden to make request</response>
 [Authorize]
 [ApiController]
 [ProducesResponseType(StatusCodes.Status200OK)]
@@ -22,16 +30,17 @@ namespace Diary.Api.Controllers;
 [ProducesResponseType(StatusCodes.Status401Unauthorized)]
 [ApiVersion("1.0")]
 [Route("api/v{version:apiVersion}/[controller]")]
-
-public class ReportController:ControllerBase
+public class ReportController : ControllerBase
 {
     private readonly IReportService _reportService;
+    private readonly IBaseRepository<Report> _reportRepository;
 
-    public ReportController(IReportService reportService)
+    public ReportController(IReportService reportService, IBaseRepository<Report> reportRepository)
     {
         _reportService = reportService;
+        _reportRepository = reportRepository;
     }
-    
+
     /// <summary>
     /// Get reports of user
     /// </summary>
@@ -45,18 +54,22 @@ public class ReportController:ControllerBase
     ///         
     ///     }
     /// </remarks>
-
-    [HttpGet("reports/{userId:int:min(0)}")]//":int:min(0)" added by myself
+    [HttpGet("reports/{userId:int:min(0)}")] //":int:min(0)" added by myself
     public async Task<ActionResult<BaseResult<ReportDto>>> GetUserReports(long userId)
     {
+        if (!CheckIsUserAllowedToGetData(userId.ToString()))
+        {
+            return Forbid("Bearer");
+        }
         var response = await _reportService.GetReportsAsync(userId);
         if (response.IsSuccess)
         {
             return Ok(response);
         }
+
         return BadRequest(response);
     }
-    
+
     /// <summary>
     /// Gets report by id
     /// </summary>
@@ -70,18 +83,22 @@ public class ReportController:ControllerBase
     ///         
     ///     }
     /// </remarks>
-
-    [HttpGet("{id:int:min(0)}")]//":int:min(0)" added by myself
+    [HttpGet("{id:int:min(0)}")] //":int:min(0)" added by myself
     public async Task<ActionResult<BaseResult<ReportDto>>> GetReport(long id)
     {
+        if(!CheckIsAnyDataBelongsToUser(id))
+        {
+            return Forbid("Bearer");
+        }
         var response = await _reportService.GetReportByIdAsync(id);
         if (response.IsSuccess)
         {
             return Ok(response);
         }
+
         return BadRequest(response);
     }
-    
+
     /// <summary>
     /// Deletes report
     /// </summary>
@@ -94,18 +111,22 @@ public class ReportController:ControllerBase
     ///         "id":1
     ///     }
     /// </remarks>
-
     [HttpDelete("{id:int:min(1)}")]
     public async Task<ActionResult<BaseResult<ReportDto>>> Delete(long id)
     {
+        if(!CheckIsAnyDataBelongsToUser(id))
+        {
+            return Forbid("Bearer");
+        }
         var response = await _reportService.DeleteReportAsync(id);
         if (response.IsSuccess)
         {
             return Ok(response);
         }
+
         return BadRequest(response);
     }
-    
+
     /// <summary>
     /// Creates Report
     /// </summary>
@@ -120,18 +141,23 @@ public class ReportController:ControllerBase
     ///         "userId":1
     ///     }
     /// </remarks>
-
     [HttpPost]
-    public async Task<ActionResult<BaseResult<ReportDto>>> Create([FromBody]CreateReportDto dto)
+    public async Task<ActionResult<BaseResult<ReportDto>>> Create([FromBody] CreateReportDto dto)
     {
+        if (!CheckIsUserAllowedToGetData(dto.UserId.ToString()))
+        {
+            return Forbid("Bearer");
+        }
+
         var response = await _reportService.CreateReportAsync(dto);
         if (response.IsSuccess)
         {
             return Ok(response);
         }
+
         return BadRequest(response);
     }
-    
+
     /// <summary>
     /// Updates report
     /// </summary>
@@ -148,14 +174,46 @@ public class ReportController:ControllerBase
     ///     }
     /// </remarks>
     [HttpPut]
-    public async Task<ActionResult<BaseResult<ReportDto>>> Update([FromBody]UpdateReportDto dto)
+    public async Task<ActionResult<BaseResult<ReportDto>>> Update([FromBody] UpdateReportDto dto)
     {
+        if(!CheckIsAnyDataBelongsToUser(dto.Id))
+        {
+            return Forbid("Bearer");
+        }
         var response = await _reportService.UpdateReportAsync(dto);
         if (response.IsSuccess)
         {
             return Ok(response);
         }
+
         return BadRequest(response);
     }
-    
+
+    private bool CheckIsUserAllowedToGetData(string identifier)
+    {
+        var userRole = User.FindFirstValue(ClaimTypes.Role);
+        List<string> canGetAnyDataRoles = new List<string>
+        {
+            nameof(Roles.Admin),
+            nameof(Roles.Moderator)
+        };
+        var canGetAnyData = canGetAnyDataRoles.Any(role=>role==userRole);
+        var id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        var isSame = id == identifier;
+        return canGetAnyData || isSame;
+    }
+
+    private bool CheckIsAnyDataBelongsToUser(long id)
+    {
+        var userRole = User.FindFirstValue(ClaimTypes.Role);
+        List<string> canGetAnyDataRoles = new List<string>
+        {
+            nameof(Roles.Admin),
+            nameof(Roles.Moderator)
+        };
+        var canGetAnyData = canGetAnyDataRoles.Any(role=>role==userRole);
+        var userId = long.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+        var userReports = _reportRepository.GetAll().AsEnumerable().Where(x => x.UserId == userId);
+        return canGetAnyData||userReports.Any(x => x.Id == id);
+    }
 }
